requirements.txt
cycler==0.10.0
dataclasses==0.8
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.2.52
Pillow==8.2.0
pydantic==1.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-stdnum==1.16
six==1.16.0
torch==1.8.1+cu111
torchvision==0.9.1+cu111
typing-extensions==3.10.0.0
